Move FileService to :feature-chat
